+++ backend/reitera-backend/src/main/java/com/helderruiz/reitera_backend/modules/user/dto/UserResponseDTO.java
@@ -9,6 +9,6 @@ public record UserResponseDTO(
         String email,
         String firstName,
         String lastName,
-        String roleName,
+        String avatarId,
         LocalDateTime createdAt
 ) {}